Update README.md
According to a current trend, it has been detected that out of every 10 applicants appearing for the job of data scientist, only 1-2 candidates get the job. So, what about the rest? They get rejected!

So, for the applicant, this is the Topmost comprehensive interview questions and answers. This will help interviewer and applicant for the hiring data/scientist and hired as data scientist/analyst.



1. What is data science? How would you say it is alike or different to corporate analytics and business aptitude?
Data science is a field that deals with an analysis of data. It studies the source of evidence, what the information epitomizes and turning it into a valuable resource by giving insights of the data that are later used for creating strategies. It is a combination of business perspectives, computer programming and statistical techniques.

Business analytics or simply analytics is the core of business intellect and data analyst. Data science is a relatively new term used for analysis of big data and giving insights.

Analytics generally has higher degree of business perspectives than data science which is more programming heavy. The terms are however used interchangeably.

2. Which package is used to do data import in R and Python? How do you do data importation in SAS?

We can do data import using multiple methods:

•	In R we use RODBC for RDBMS data, and data.table for fast import.
•	We use jsonlite for JSON data, foreign package for other languages like SPSS
•	We use data and sas7bdat package for SAS data.
•	In Python we use Pandas package and the commands read_csv , read_sql for reading data. Also, we can use SQL Alchemy in Python for connecting to databases.

3. What is an RDBMS? Name some examples for RDBMS? What is CRUD?

A relational database management system (RDBMS) is a database management system that is based on a relational model. The relational prototypical uses the basic concept of a relation or table. RDBMS is the basis for SQL, and for database systems like MS SQL Server, IBM DB2, Oracle, MySQL, and Microsoft Access.

In computer programming, create, read, update and delete (as an acronym CRUD or possibly a backronym) (Sometimes called SCRUD with an “S” for Search) are the four basic functions of persistent storage.

4. What is JSON and What is XML?

JSON is an condensation for JavaScript Object Notation. It is a primary data format that uses human-readable text to transfer data objects consisting of data interpretation language. Although originally derived from the JavaScript scripting language, JSON is a language-independent data format. Code for generating JSON data is readily available in many programming languages.

XML is an abbreviation for Extensible Markup Language. It defines a set of rules that is used for encoding documents in a human and machine-readable format. The design goals of XML emphasize simplicity, generality and usability across the Internet. It is a documented data format with strong support for different human languages. It is mostly used for the representation of uninformed data structures such as those used in web services.

5. How do you check for data quality?

Data quality is a valuation of data’s suitability to serve its purpose in a given context. Different aspects of data quality include:

•	Accuracy
•	Completeness
•	Update status
•	Relevance
•	Consistency across data sources
•	Reliability
•	Appropriate presentation
•	Accessibility

Keeping data quality needs going through the data in different intervals and scrubbing it. This involves informing it, normalizing it, and removing copies to generate a single view of the data, even if it is stored in multiple systems.
For more click here: 
https://www.econolytics.in/blog/data-analyst-scientist-interview-questions-answers/